+++ CHANGELOG.rst
@@ -7,6 +7,12 @@ Change Log
 ----------
 
 
+1.4.0
+=====
+* 2025-01-15 / branch: dmichaels-custom-column-mappings-20250115 / PR-16 / dmichaels
+* Added custom column mappings for simplified QCs specification in spreadsheet. 
+
+
 1.3.0
 =====
 * Fix for C4-1187 where we were not implicitly adding consortia to submitted like for submission-centers.